cleanup: isolate groove terminal delta handling
